feature: add :core:mock
