+++ feature/edit/src/test/java/kosenda/makecolor/feature/edit/RegisterViewModelImplTest.kt
@@ -2,12 +2,10 @@ package kosenda.makecolor.feature.edit
 
 import androidx.lifecycle.SavedStateHandle
 import com.google.common.truth.Truth.assertThat
-import kosenda.makecolor.core.data.repository.ColorRepository
+import kosenda.makecolor.core.mock.FakeColorRepository
 import kosenda.makecolor.core.model.data.Category
-import kosenda.makecolor.core.model.data.ColorItem
 import kosenda.makecolor.core.testing.MainDispatcherRule
 import kosenda.makecolor.core.ui.data.NavKey
-import kosenda.makecolor.core.util.convertDisplayStringListFromCategories
 import kosenda.makecolor.core.util.randomColorData
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.serialization.encodeToString
@@ -20,6 +18,7 @@ class RegisterViewModelImplTest {
 
     @get: Rule
     val mainDispatcherRule = MainDispatcherRule()
+
     private val initialColorDataJson = Json.encodeToString(randomColorData())
     private val registerViewModel = RegisterViewModelImpl(
         ioDispatcher = mainDispatcherRule.testDispatcher,
@@ -62,41 +61,3 @@ class RegisterViewModelImplTest {
     }
 }
 
-class FakeColorRepository : ColorRepository {
-
-    private val colors = mutableListOf<ColorItem>()
-    private val categories = mutableListOf(
-        Category("test1", 0),
-        Category("test2", 0),
-    )
-
-    override suspend fun getColors(category: String): List<ColorItem> {
-        return colors.filter { it.category == category }
-    }
-    override suspend fun getSize(category: String): Int = colors.size
-    override suspend fun isExistCategory(category: String): Boolean {
-        if (categories.isEmpty()) return false
-        convertDisplayStringListFromCategories(categories).forEach {
-            if (it == category) return true
-        }
-        return false
-    }
-    override suspend fun updateSize(size: Int, name: String) {}
-    override suspend fun updateCategory(oldName: String, newName: String) {}
-    override suspend fun colorChangeCategory(category: String, newCategory: String) {}
-    override suspend fun deleteColor(colorItem: ColorItem) {}
-    override suspend fun decreaseSize(name: String) {}
-    override suspend fun deleteCategory(name: String) {}
-    override suspend fun deleteColors(category: String) {}
-    override suspend fun deleteAllCategories() {}
-    override suspend fun deleteAllColors() {}
-    override suspend fun insertColor(color: ColorItem) {
-        colors.add(color)
-    }
-    override suspend fun getCategory(): List<Category> {
-        return categories
-    }
-    override suspend fun insertCategory(category: Category) {
-        categories.add(category)
-    }
-}
